+++ apps/connect/source/ftrack_connect/plugin_manager/plugin_list.py
@@ -239,6 +239,19 @@ class DndPluginList(QtWidgets.QFrame):
                 return item
         return None
 
+    def _remove_plugin(self, plugin_name):
+        '''Remove the plugin *plugin_name* from plugin list (not disk),
+        if succeeded/found True will be returned, False otherwise.'''
+        num_items = self._plugin_model.rowCount()
+        for i in range(num_items):
+            item = self._plugin_model.item(i)
+            item_name = item.data(ROLES.PLUGIN_NAME)
+            if item_name == plugin_name:
+                # Remove item
+                self._plugin_model.takeRow(i)
+                return True
+        return False
+
     def populate_installed_plugins(self, plugins):
         '''Populate model with installed plugins.'''
         self._installed_plugin_count = 0
@@ -364,7 +377,10 @@ class DndPluginList(QtWidgets.QFrame):
         paths = self._process_mime_data(event.mimeData())
 
         for path in paths:
-            self._add_plugin(get_plugin_data(path), STATUSES.NEW)
+            # Remove existing one
+            plugin_data = get_plugin_data(path)
+            self._remove_plugin(plugin_data['name'])
+            self._add_plugin(plugin_data, STATUSES.NEW)
 
         event.accept()
         self._set_drop_zone_state()