+++ libs/qt/source/ftrack_qt/widgets/frames/__init__.py
@@ -1,7 +1,0 @@
-from ftrack_qt.widgets.frames.asset_version_creation import (
-    AssetVersionCreation,
-)
-from ftrack_qt.widgets.frames.asset_version_selection import (
-    AssetVersionSelection,
-)
-from ftrack_qt.widgets.frames.new_asset_input import NewAssetInput